requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.5
Django==4.2
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
joblib==1.2.0
Levenshtein==0.21.0
nltk==3.8.1
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pycparser==2.21
pyotp==2.8.0
PySocks==1.7.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.0
pytz==2023.3
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.31.0
scikit-surprise==1.1.3
scipy==1.10.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
thefuzz==0.19.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
